refactor: refactor my-poor-room
