--- src/my-poor-room/components/achievement-of-challenge-information/index.ts
@@ -1,3 +1,0 @@
-import { AchievementOfChallengeInformation } from './AchievementOfChallengeInformation';
-
-export { AchievementOfChallengeInformation };
