+++ src/my-poor-room/components/challenge-achievement/index.ts
@@ -1,0 +1,4 @@
+import { ChallengeAchievement } from './ChallengeAchievement';
+import { ProgressBar } from './Progressbar';
+
+export { ChallengeAchievement, ProgressBar };
